pyproject.toml
[project]
name = "review-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Mostafa Hassan"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
